Update pins_BTT_OCTOPUS_V1_common.h (#75)
Introduced a (quick) fix for allowing use of MKS12864A/B and MKS_MINI_12864 LCD. See issue #70. Credits for the original work to @Gleb1982
The only small issue remaining is that FPC cables will need to be inverted (rotated by 180°) when connecting to the motherboard socket. Even tough, it's a trivial issue that mostly are familiar with.