Fix to the absurdly low max_z_accel
I have been helping a few people with their printers and keep finding issues relating to this config. The max_z_accel set here is absurdly low, and is so slow that it causes blobs for whenever a layer change is requested by gcode.
I have updated it with a more appropriate acceleration, that should be suitable for a leadscrew based z axis.
Although it's good practise for people to completely customise their own configurations, not everyone is as experienced as others, and small little settings like these can cause extra unneccesary confusion and troubleshooting.